[meaty_skeleton] reorganized source tree to follow wiki.osdev.org recommended layout for the meaty skeleton operating system
