+++ build.sh
@@ -3,5 +3,5 @@ set -e
 . ./headers.sh
 
 for PROJECT in $PROJECTS; do
-	DESTDIR="$PWD/sysroot" $MAKE -C $PROJECT install
+	(cd $PROJECT && DESTDIR="$SYSROOT" $MAKE install)
 done
