+++ clean.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+set -e
+. ./config.sh
+
+for PROJECT in $PROJECTS; do
+	(cd $PROJECT && $MAKE clean)
+done
+
+rm -rf sysroot
+rm -rf isodir
+rm -rf missy.iso
