+++ config.sh
@@ -1,0 +1,27 @@
+SYSTEM_HEADER_PROJECTS="libc kernel"
+PROJECTS="libc kernel"
+
+export MAKE=${MAKE:-make}
+export HOST=${HOST:-$(./default-host.sh)}
+
+export AR=${HOST}-ar
+export AS=${HOST}-as
+export CC=${HOST}-gcc
+
+export PREFIX=/usr
+export EXEC_PREFIX=$PREFIX
+export BOOTDIR=/boot
+export LIBDIR=$EXEC_PREFIX/lib
+export INCLUDEDIR=$PREFIX/include
+
+export CFLAGS='-O2 -g'
+export CPPFLAGS=''
+
+# use desired OS sysroot
+export SYSROOT="$(pwd)/sysroot"
+export CC="$CC --sysroot=$SYSROOT"
+
+# workaround for --without-headers in the early stages
+if echo "$HOST" | grep -Eq -- '-elf($|-)'; then
+	export CC="$CC -isystem=$INCLUDEDIR"
+fi
