+++ init/arch/bootstub.asm
