+++ init/include/kterm.h
