+++ init/include/kutils.h
