+++ init/make.config
@@ -1,0 +1,15 @@
+KERNEL_INIT_CFLAGS=
+KERNEL_INIT_CPPFLAGS=
+KERNEL_INIT_LDFLAGS=
+KERNEL_INIT_LIBS=
+
+KERNEL_INIT_OBJS=\
+	$(INITDIR)/init/kinit.o \
+	$(INITDIR)/init/kterm.o \
+	$(INITDIR)/init/kutils.o \
+	$(INITDIR)/init/multiboot/api.o \
+	$(INITDIR)/init/multiboot/mboot.o \
+	$(INITDIR)/init/multiboot/mboot2.o \
+	$(INITDIR)/init/memory/kmemlow.o \
+	$(INITDIR)/init/memory/paging.o \
+	$(INITDIR)/arch/bootstub.o
