+++ kernel/arch/i386/make.config
@@ -1,0 +1,7 @@
+KERNEL_ARCH_CFLAGS=
+KERNEL_ARCH_CPPFLAGS=
+KERNEL_ARCH_LDFLAGS=
+KERNEL_ARCH_LIBS=
+
+KERNEL_ARCH_OBJS=\
+	$(ARCHDIR)/display.o
