+++ libc/Makefile
@@ -1,0 +1,89 @@
+#
+# MISSY Microsystem kernel makefile
+# Outline taken from OSDev.org Meaty Skeleton project, modified to meet the needs
+# of MISSY Microsystem Alpha 3.0 build
+#
+# Ben Melikant, 8/2016
+#
+DEFAULT_HOST!=../default-host.sh
+HOST?=DEFAULT_HOST
+HOSTARCH!=../target-triplet-to-arch.sh $(HOST)
+
+CFLAGS?=-O2 -g
+CPPFLAGS?=
+LDFLAGS?=
+LIBS?=
+
+DESTDIR?=
+PREFIX?=/usr/local
+EXEC_PREFIX?=$(PREFIX)
+INCLUDEDIR?=$(PREFIX)/include
+LIBDIR?=$(EXEC_PREFIX)/lib
+
+CFLAGS:=$(CFLAGS) -ffreestanding -fbuiltin -Wall -Wextra
+CPPFLAGS:=$(CPPFLAGS) -D__is_missy_kernel -DDEBUG_BUILD -D__build_i386 -Iinclude -Iarch -Ikernel -DDEBUG_KMEMLOW 
+LIBK_CFLAGS:=$(CFLAGS)
+LIBK_CPPFLAGS:=$(CPPFLAGS) -D__is_libk
+
+ARCHDIR=arch/$(HOSTARCH)
+include $(ARCHDIR)/make.config
+
+CFLAGS:=$(CFLAGS) $(ARCH_CFLAGS)
+CPPFLAGS:=$(CPPFLAGS) $(ARCH_CPPFLAGS)
+LIBK_CFLAGS:=$(LIBK_CFLAGS) $(KERNEL_ARCH_CFLAGS)
+LIBK_CPPFLAGS:=$(LIBK_CPPFLAGS) $(KERNEL_ARCH_CPPFLAGS)
+
+FREEOBJS=\
+	$(ARCH_FREEOBJS)
+
+HOSTEDOBJS=\
+	$(ARCH_HOSTEDOBJS)
+
+OBJS=\
+	$(FREEOBJS) \
+	$(HOSTEDOBJS) \
+
+LIBK_OBJS=$(FREEOBJS:.o=.libk.o)
+
+BINARIES=libk.a #libc.a not ready yet
+
+.PHONY: all clean install install-headers install-libs
+.SUFFIXES: .o .libk.o .c .S
+
+all: $(BINARIES)
+
+libc.a: $(OBJS)
+	$(AR) rcs $@ $(OBJS)
+
+libk.a: $(LIBK_OBJS)
+	$(AR) rcs $@ $(LIBK_OBJS)
+
+.c.o:
+	$(CC) -MD -c $< -o $@ -std=gnu11 $(CFLAGS) $(CPPFLAGS)
+
+.c.S:
+	$(CC) -MD -c $< -o $@ $(CFLAGS) $(CPPFLAGS)
+
+.c.libk.o:
+	$(CC) -MD -c $< -o $@ -std=gnu11 $(LIBK_CFLAGS) $(LIBK_CPPFLAGS)
+
+.S.libk.o:
+	$(CC) -MD -c $< -o $@ $(LIBK_CFLAGS) $(LIBK_CPPFLAGS)
+
+clean:
+	rm -f $(BINARIES) *.a
+	rm -f $(OBJS) $(LIBK_OBJS) *.o */*.o */*/*.o
+	rm -f $(OBJS:.o=.d) $(LIBK_OBJS:.o=.d) *.d */*.d */*/*.d
+
+install: install-headers install-libs
+
+install-headers:
+	mkdir -p $(DESTDIR)$(INCLUDEDIR)
+	cp -R --preserve=timestamps include/. $(DESTDIR)$(INCLUDEDIR)/.
+
+install-libs: $(BINARIES)
+	mkdir -p $(DESTDIR)$(LIBDIR)
+	cp $(BINARIES) $(DESTDIR)$(LIBDIR)
+
+-include $(OBJS:.o=.d)
+-include $(LIBK_OBJS:.o=.d)
