+++ libc/arch/i386/make.config
@@ -1,0 +1,8 @@
+ARCH_CFLAGS=
+ARCH_CPPFLAGS=
+KERNEL_ARCH_CFLAGS=
+KERNEL_ARCH_CPPFLAGS=
+
+ARCH_FREEOBJS=\
+
+ARCH_HOSTEDOBJS=\
