--- src/kernel/kernel/kterminal.c
@@ -1,7 +1,0 @@
-#include <kernel/kterminal.h>
-
-#ifdef __build_i386
-#include <kernel/driver/i386/display.h>
-#include <kernel/driver/i386/kbinput.h>
-#endif
-
